+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Privacy API: user data export now includes the user's flags and the files embedded in their reviews;
   deleting user data also deletes these files and the flags on the deleted reviews
 - Images and files embedded in reviews are no longer accessible to users who are not allowed to see the reviews
+- Users can no longer flag ratings in courses where they are not allowed to see the reviews
 
 ## [4.5.1] - 2026-04-17
 ### Fixed